+++ shadowsocks/config.go
@@ -38,6 +38,7 @@ type Config struct {
 }
 
 var readTimeout time.Duration
+var udpTimeout time.Duration
 
 func (config *Config) GetServerArray() []string {
 	// Specifying multiple servers in the "server" options is deprecated.
@@ -91,6 +92,7 @@ func ParseConfig(path string) (config *Config, err error) {
 		config.Method = config.Method[:len(config.Method)-5]
 		config.Auth = true
 	}
+	udpTimeout = time.Duration(config.Timeout) * time.Second
 	return
 }
 
@@ -133,4 +135,5 @@ func UpdateConfig(old, new *Config) {
 
 	old.Timeout = new.Timeout
 	readTimeout = time.Duration(old.Timeout) * time.Second
+	udpTimeout = time.Duration(old.Timeout) * time.Second
 }
